requirements.txt
python_full_version>=3.7.0b1
python-telegram-bot>=12.5.1
